very simple refactoring
